requirements.txt
gspread
selenium 
webdriver-manager
undetected-chromedriver
google-auth 
bs4
pandas
fake_useragent
PyYAML
streamlit


-e .